Atualização Faturamento e pedido

+++ src/java/br/sk8line/dao/PedidoDAO.java
@@ -1,7 +1,9 @@
 package br.sk8line.dao;
 
 import br.sk8line.modelo.Pedido;
+import java.util.List;
 import javax.persistence.EntityManager;
+import javax.persistence.Query;
 
 public class PedidoDAO {
     
@@ -29,8 +31,18 @@ public class PedidoDAO {
         return em.find(Pedido.class, id);
     }
     
-    public void remover(Long id){
+    public void remover(Long id)throws Exception{
         Pedido p = consultarPorId(id);
+        
+        if (p.getStatus().equals("F")){
+            throw new Exception("Pedido já faturado! Não pode ser excluído!");
+        }
+        
         em.remove(p);
+    }
+    
+    public List<Pedido> consultarTodos(){
+        Query consulta = em.createNamedQuery("Pedido.consultarTodos");
+        return consulta.getResultList();
     }
 }
